Changed double to int

+++ Program.cs
@@ -7,11 +7,11 @@ namespace First_CSharp_Program
         static void Main(string[] args)
         {
             string name;
-            double age;
+            int age;
             Console.WriteLine("Enter your name: ");
             name = Console.ReadLine();
             Console.WriteLine("Enter your age: ");
-            age = Convert.ToDouble(Console.ReadLine());
+            age = Convert.ToInt32(Console.ReadLine());
             Console.WriteLine($"{name}, next year you will be {age + 1}");
 
         }
